chore: upgrade webpack to fix ssl3 errors
When attempting to run the build in node versions >= 17, this error was
reported when running webpack: `Error: error:0308010C:digital envelope
routines::unsupported`. This was due to Node 17 using SSL3, breaking
features of Webpack that use it. The upgrade to latest Webpack fixes
the issue since newer version leverage a wasm md4 hashing strategy.